Scaling Fix
Game didn't scale properly when screen size is changed. This was due to using the window size instead of the constant size.

The constant size can be used as the window size and the game window size are two separate values and the game window size is the same constant value that aligns with position and other values used.