CONSTANTS Formatting
Updated all the files to use the CONSTANTS values.

So it appeared that across the files a lot of them were ignoring the CONSTANT values even if they existed and even some made their own local CONSTANTS values that already existed.

Pretty much just moved all constant values to CONSTANTS and updated any hard coded values to be in CONSTANTS